separated importable tool libraries from command line scripts
