+++ .gitignore
@@ -31,5 +31,8 @@ configs/test_execution.conf
 
 ### Tutorial created files ###
 Tutorials/01_introduction/custom_config.conf
-
+Tutorials/03_fixed_station_data/ostergarnsholm_reanalysed.conf
+Tutorials/03_fixed_station_data/ostergarnsholm_original.conf
+Tutorials/04_N2O_fluxes/N2O_rapid_N00.conf
+Tutorials/04_N2O_fluxes/N2O_rapid_N00_surfactants.conf
