+++ core/rate_parameterisation.py
