+++ core/settings.xml
