+++ scripts/reanalyse_socat_driver.py
@@ -2,15 +2,9 @@
 # -*- coding: utf-8 -*-
 """
 Created on Mon Jul  2 13:44:02 2018
-
+Drives the reanalyse_socat tool and provides access via commandline.
 @author: Tom Holding
 """
-
-import inspect;
-import argparse;
-from os import path, makedirs;
-import shutil; ##move desired output from reanalyse_socat to the specified output folder and delete undesired output
-
 
 #Example options:
 #Single small region test run, ascii output, no coastal file.
@@ -45,8 +39,12 @@ import shutil; ##move desired output from reanalyse_socat to the specified outpu
 #Whole global SOCATv6 reanalysis (ASCII)
 #   -socat_dir ~/data/SOCAT_ascii/SOCATv6/ -socat_files SOCATv6.tsv -sst_dir ~/data/ocean_flux_ftp/SST_reynolds_avhrr/ -sst_tail 01_OCF-SST-GLO-1M-100-REYNOLDS.nc -output_dir ~/Files/fluxengine_v3/FluxEngine/output/reanalysis_socat_output_gridded -socatversion 4 -usereynolds -startyr 1957 -endyr 2017 -keepduplicates -keeptempfiles -asciioutput
 
+import inspect;
+import argparse;
+from os import path, makedirs;
+import shutil; #move desired output from reanalyse_socat to the specified output folder and delete undesired output
 
-from . import reanalyse_socat.reanalyse_socat_v2 as rs;
+import fluxengine.tools.reanalyse_socat.reanalyse_socat_v2 as rs;
 
 
 if __name__ == "__main__":
