+++ scripts/text2ncdf_examples.sh
