+++ tools/lib_compare_net_budgets.py
@@ -7,7 +7,6 @@ Created on Wed Apr 11 10:50:09 2018
 """
 
 from numpy import loadtxt;
-from sys import argv;
 
 #read the global flux budgets output.
 def read_global_core_budgets(outputPath):
@@ -37,15 +36,3 @@ def calc_net_budget_percentages(outputPath, referencePath, verbose=False):
         percentages[colName] = percent;
     
     return percentages;
-
-
-
-if __name__ == "__main__":
-    if len(argv != 3):
-        print("Usage: compare_net_budgets newOutputPath referenceOutputPath");
-    else:
-        outputPath = argv[1];
-        referencePath = argv[2];    
-        calc_net_budget_percentages(outputPath, referencePath);
-
-
