+++ tools/reanalyse_socat/commands_to_run_to_concat_reanalysed_ascii.txt
